Give user full control over state management
You can imagine situations where the state transition logic
provided by the library doesn't give a developer the flexibility they
need. Maybe their user flow includes other steps between selecting the
reason and submitting it, or maybe they want to initialize the
component in a state other than `Unselected`

In general, in any situation where the component's state depends on some
 other application state, it makes sense to provide the component with a
"config-only" interface: all the state is exposed, so the developer
has the flexibility to pick out which state render, while still not
having to worry about how that config/state is turned into Html

To get the best of both worlds, you can reimplement the same "statefull"
interface in terms of this "config-only" one, and expose both as
separate modules. I'll sketch that out in code when I get a chance